backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
aiohttp==3.9.1
celery==5.3.4
redis==5.0.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
networkx==3.2.1
sentence-transformers==2.2.2
loguru==0.7.2
prometheus-client==0.19.0
fastapi-limiter==0.1.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
perplexity-ai==0.2.3